# packages/api-client/src/openapi.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
...
  /api/v1/recommendations/{id}:
    put:
      tags:
      - recommendation-controller
      operationId: update
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRecommendationRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
    delete:
      tags:
      - recommendation-controller
      operationId: delete
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
...
  /api/v1/companies/me:
    get:
      tags:
      - company-controller
      operationId: getMine_1
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseCompanyResponse"
    put:
      tags:
      - company-controller
      operationId: update_6
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCompanyRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
    delete:
      tags:
      - company-controller
      operationId: delete_4
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
...
  /api/v1/recommendations:
    post:
      tags:
      - recommendation-controller
      operationId: create
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRecommendationRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseLong"
  /api/v1/recommendations/{id}/show:
    post:
      tags:
      - recommendation-controller
      operationId: show
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
  /api/v1/recommendations/{id}/hide:
    post:
      tags:
      - recommendation-controller
      operationId: hide
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
...
  /api/v1/companies:
    get:
      tags:
      - company-controller
      operationId: list_7
      parameters:
      - name: cursorLimit
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/CursorLimit"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseCursorPageCompanyResponse"
    post:
      tags:
      - company-controller
      operationId: create_11
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCompanyRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseLong"
  /api/v1/companies/{id}/deny:
    post:
      tags:
      - company-controller
      operationId: deny_3
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
  /api/v1/companies/{id}/accept:
    post:
      tags:
      - company-controller
      operationId: accept_3
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
...
  /api/v1/recommendations/sent:
    get:
      tags:
      - recommendation-controller
      operationId: listSent
      parameters:
      - name: memberId
        in: query
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseListRecommendationResponse"
  /api/v1/recommendations/received:
    get:
      tags:
      - recommendation-controller
      operationId: listReceived
      parameters:
      - name: memberId
        in: query
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseListRecommendationResponse"
...
  /api/v1/companies/{id}:
    get:
      tags:
      - company-controller
      operationId: get_8
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseCompanyResponse"
  /api/v1/companies/{id}/members:
    get:
      tags:
      - company-controller
      operationId: listMembers
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseListMemberSummaryResponse"
...
components:
  schemas:
    Address:
      type: object
      properties:
        zipcode:
          type: string
          minLength: 1
          pattern: "\\d{5}"
        bcode:
          type: string
          minLength: 1
          pattern: "\\d{10}"
        state:
          $ref: "#/components/schemas/Region"
        city:
          type: string
          minLength: 1
        street:
          type: string
          minLength: 1
        detail:
          type: string
        latitude:
          type: number
        longitude:
          type: number
      required:
      - bcode
      - city
      - latitude
      - longitude
      - state
      - street
      - zipcode
    Region:
      type: string
      enum:
      - 서울
      - 부산
      - 대구
      - 인천
      - 대전
      - 울산
      - 세종특별자치시
      - 경기
      - 강원특별자치도
      - 충북
      - 충남
      - 전북특별자치도
      - 전남광주통합특별시
      - 경북
      - 경남
      - 제주특별자치도
...
    Trade:
      type: string
      enum:
      - DESIGN
      - DEMOLITION
      - ELECTRICAL
      - PLUMBING
      - MECHANICAL
      - MASONRY
      - CARPENTRY
      - GLAZING
      - WATERPROOFING
      - PLASTERING
      - INSULATION
      - TILING
      - GROUTING
      - PAINTING
      - WALLPAPER
      - FILM_SHEET
      - HARDWOOD
      - VINYL
      - SINK
      - FURNITURE
      - AIR_CONDITIONING
      - HOISTING
      - TRANSPORT
      - CLEANING
      - GENERAL_LABOR
      - HOME_REPAIR
      - BATHROOM
      - DISINFECTION
...
    ApiResponseVoid:
      type: object
      properties:
        success:
          type: boolean
        error:
          $ref: "#/components/schemas/ExceptionCode"
        data: {}
    ExceptionCode:
      type: object
      properties:
        message:
          type: string
        status:
          $ref: "#/components/schemas/HttpStatus"
        code:
          type: string
    HttpStatus:
      type: string
      enum:
      - 100 CONTINUE
      - 101 SWITCHING_PROTOCOLS
      - 102 PROCESSING
      - 103 EARLY_HINTS
      - 200 OK
      - 201 CREATED
      - 202 ACCEPTED
      - 203 NON_AUTHORITATIVE_INFORMATION
      - 204 NO_CONTENT
      - 205 RESET_CONTENT
      - 206 PARTIAL_CONTENT
      - 207 MULTI_STATUS
      - 208 ALREADY_REPORTED
      - 226 IM_USED
      - 300 MULTIPLE_CHOICES
      - 301 MOVED_PERMANENTLY
      - 302 FOUND
      - 303 SEE_OTHER
      - 304 NOT_MODIFIED
      - 307 TEMPORARY_REDIRECT
      - 308 PERMANENT_REDIRECT
      - 400 BAD_REQUEST
      - 401 UNAUTHORIZED
      - 402 PAYMENT_REQUIRED
      - 403 FORBIDDEN
      - 404 NOT_FOUND
      - 405 METHOD_NOT_ALLOWED
      - 406 NOT_ACCEPTABLE
      - 407 PROXY_AUTHENTICATION_REQUIRED
      - 408 REQUEST_TIMEOUT
      - 409 CONFLICT
      - 410 GONE
      - 411 LENGTH_REQUIRED
      - 412 PRECONDITION_FAILED
      - 413 CONTENT_TOO_LARGE
      - 413 PAYLOAD_TOO_LARGE
      - 414 URI_TOO_LONG
      - 415 UNSUPPORTED_MEDIA_TYPE
      - 416 REQUESTED_RANGE_NOT_SATISFIABLE
      - 417 EXPECTATION_FAILED
      - 418 I_AM_A_TEAPOT
      - 421 MISDIRECTED_REQUEST
      - 422 UNPROCESSABLE_CONTENT
      - 422 UNPROCESSABLE_ENTITY
      - 423 LOCKED
      - 424 FAILED_DEPENDENCY
      - 425 TOO_EARLY
      - 426 UPGRADE_REQUIRED
      - 428 PRECONDITION_REQUIRED
      - 429 TOO_MANY_REQUESTS
      - 431 REQUEST_HEADER_FIELDS_TOO_LARGE
      - 451 UNAVAILABLE_FOR_LEGAL_REASONS
      - 500 INTERNAL_SERVER_ERROR
      - 501 NOT_IMPLEMENTED
      - 502 BAD_GATEWAY
      - 503 SERVICE_UNAVAILABLE
      - 504 GATEWAY_TIMEOUT
      - 505 HTTP_VERSION_NOT_SUPPORTED
      - 506 VARIANT_ALSO_NEGOTIATES
      - 507 INSUFFICIENT_STORAGE
      - 508 LOOP_DETECTED
      - 509 BANDWIDTH_LIMIT_EXCEEDED
      - 510 NOT_EXTENDED
      - 511 NETWORK_AUTHENTICATION_REQUIRED
...
    UpdateRecommendationRequest:
      type: object
      properties:
        content:
          type: string
          minLength: 1
      required:
      - content
...
    ProfileRole:
      type: string
      enum:
      - CONTRACTOR
      - CLIENT
      - ARCHITECT
      - FOREMAN
      - SKILLED
      - SEMI_SKILLED
      - HELPER
...
    UpdateCompanyRequest:
      type: object
      properties:
        pictureId:
          type: integer
          format: int64
...
    ApiResponseLong:
      type: object
      properties:
        success:
          type: boolean
        error:
          $ref: "#/components/schemas/ExceptionCode"
        data:
          type: integer
          format: int64
...
    CreateRecommendationRequest:
      type: object
      properties:
        toId:
          type: integer
          format: int64
        content:
          type: string
          minLength: 1
      required:
      - content
      - toId
...
    CreateCompanyRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        brn:
          type: string
          minLength: 1
        pictureId:
          type: integer
          format: int64
        attachmentId:
          type: integer
          format: int64
      required:
      - attachmentId
      - brn
      - name
...
    MemberSummaryResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        name:
          type: string
        picture:
          type:
          - string
          - "null"
      required:
      - id
      - name
      - picture
      - username
...
    ProfileSummaryResponse:
      type: object
      properties:
        role:
          $ref: "#/components/schemas/ProfileRole"
        primaryTrade:
          $ref: "#/components/schemas/Trade"
        experience:
          type: integer
          format: int32
        headline:
          type:
          - string
          - "null"
        address:
          $ref: "#/components/schemas/Address"
      required:
      - address
      - experience
      - headline
      - primaryTrade
      - role
    ApiResponseListRecommendationResponse:
      type: object
      properties:
        success:
          type: boolean
        error:
          $ref: "#/components/schemas/ExceptionCode"
        data:
          type: array
          items:
            $ref: "#/components/schemas/RecommendationResponse"
    RecommendationResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        member:
          $ref: "#/components/schemas/MemberSummaryResponse"
        profile:
          $ref: "#/components/schemas/ProfileSummaryResponse"
        content:
          type: string
        visible:
          type: boolean
        createdAt:
          type: string
          format: date-time
        modifiedAt:
          type: string
          format: date-time
      required:
      - content
      - createdAt
      - id
      - member
      - modifiedAt
      - profile
      - visible
...
    ApiResponseListMemberSummaryResponse:
      type: object
      properties:
        success:
          type: boolean
        error:
          $ref: "#/components/schemas/ExceptionCode"
        data:
          type: array
          items:
            $ref: "#/components/schemas/MemberSummaryResponse"
    CursorLimit:
      type: object
      properties:
        cursor:
          type: integer
          format: int64
        limit:
          type: integer
          format: int32
        reverse:
          type: boolean
...
    ApiResponseCursorPageCompanyResponse:
      type: object
      properties:
        success:
          type: boolean
        error:
          $ref: "#/components/schemas/ExceptionCode"
        data:
          $ref: "#/components/schemas/CursorPageCompanyResponse"
    CompanyResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        memberId:
          type: integer
          format: int64
        name:
          type: string
        brn:
          type: string
        status:
          $ref: "#/components/schemas/CompanyStatus"
        picture:
          type:
          - string
          - "null"
        createdAt:
          type: string
          format: date-time
        modifiedAt:
          type: string
          format: date-time
      required:
      - brn
      - createdAt
      - id
      - memberId
      - modifiedAt
      - name
      - picture
      - status
    CompanyStatus:
      type: string
      enum:
      - PENDING
      - ACCEPTED
      - DENIED
    CursorPageCompanyResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/CompanyResponse"
        hasNext:
          type: boolean
        nextCursor:
          type: integer
          format: int64
    ApiResponseCompanyResponse:
      type: object
      properties:
        success:
          type: boolean
        error:
          $ref: "#/components/schemas/ExceptionCode"
        data:
          $ref: "#/components/schemas/CompanyResponse"
...